#9 webpack

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 export const home = async (req, res) => {
   try {
@@ -10,16 +11,21 @@ export const home = async (req, res) => {
 };
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
   if (!video) return res.render("404", { pageTitle: "Video not found." });
-
   return res.render("watch", { pageTitle: video.title, video });
 };
 
+///////////////////////
+
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
   if (!video) return res.render("404", { pageTitle: "Video not found" });
+  if (String(video.owner) !== String(_id)) return res.status(403).redirect("/");
   return res.render("edit", { pageTitle: `Editing`, video });
 };
 
@@ -35,26 +41,50 @@ export const postEdit = async (req, res) => {
   });
   return res.redirect(`/videos/${id}`);
 };
+///////////////////
 
 export const getUpload = (req, res) => {
   return res.render("upload", { pageTitle: "Upload Video" });
 };
 
 export const postUpload = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
+  const { path: fileUrl } = req.file;
   const { title, description, hashtags } = req.body;
-  await Video.create({
+  const newVideo = await Video.create({
     title,
     description,
+    fileUrl,
+    owner: _id,
     hashtags: Video.formatHashtags(hashtags),
   });
+  const user = await User.findById(_id);
+  user.videos.push(newVideo._id);
+  user.save();
   return res.redirect("/");
 };
 
+///////////////
+
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+  const user = await User.findById(_id);
+  if (!video)
+    return res.status(404).render("404", { pageTitle: "Video not found." });
+  if (String(video.owner) !== String(_id)) return res.status(403).redirect("/");
   await Video.findByIdAndDelete(id);
+  user.videos.splice(user.videos.indexOf(id), 1);
+  user.save();
   return res.redirect("/");
 };
+
+////////////////
 
 export const search = async (req, res) => {
   const { keyword } = req.query;
